Add tests for preventing selection of an occupied seat
